feat: separated package logic for server, client and shared
Since the minizinc library caused errors when importing Pricing4TS in client apps, the library has been refactored to only export client-friendly components through its main import, i.e. , and to export the server-oriented components from the server export, i.e.
