+++ package.json
@@ -1,9 +1,19 @@
 {
   "name": "pricing4ts",
-  "version": "0.3.2",
+  "version": "0.4.0",
   "description": "![NPM Version](https://img.shields.io/npm/v/pricing4ts)\nPricing4TS is a TypeScript-based toolkit designed to enhance the server-side functionality of a pricing-driven SaaS by enabling the seamless integration of pricing plans into the application logic. The package provides a suite of components that are predicated on the Pricing2Yaml syntax, a specification that facilitates the definition of system's pricing and its features alongside their respective evaluation expressions, grouping them within plans and add-ons, as well as establishing usage limits.",
   "main": "dist/main.js",
   "types": "dist/main.d.ts",
+  "exports": {
+    ".": {
+      "import": "./dist/main.js",
+      "require": "./dist/main.js"
+    },
+    "./server": {
+      "import": "./dist/server/server.js",
+      "require": "./dist/server/server.js"
+    }
+  },
   "directories": {
     "test": "tests"
   },
@@ -34,10 +44,12 @@
     "@types/papaparse": "^5.3.15",
     "assert": "^2.1.0",
     "js-yaml": "^4.1.0",
-    "minizinc": "^4.3.5",
     "papaparse": "^5.4.1",
     "semver": "^7.6.3",
     "uuid": "^10.0.0"
+  },
+  "peerDependencies": {
+    "minizinc": "^4.3.5"
   },
   "devDependencies": {
     "@types/assert": "^1.5.10",
