+++ src/main.ts
@@ -1,19 +1,12 @@
-import PricingService from './services/pricing.service';
-
 export {
   retrievePricingFromPath,
   retrievePricingFromYaml,
   writePricingToYaml,
-} from './utils/yaml-utils';
+} from './main/utils/yaml-utils';
 
-export { pricing2DZN } from './utils/dzn-exporter/pricing-dzn-exporter';
-export { saveDZNfile } from './utils/dzn-exporter';
-
-export { PricingService };
-
-export type { Pricing } from './models/pricing2yaml/pricing';
-export type { Feature } from './models/pricing2yaml/feature';
-export type { UsageLimit } from './models/pricing2yaml/usage-limit';
-export type { Plan } from './models/pricing2yaml/plan';
-export type { AddOn } from './models/pricing2yaml/addon';
-export type { RenderMode } from './models/pricing2yaml/types';
+export type { Pricing } from './main/models/pricing2yaml/pricing';
+export type { Feature } from './main/models/pricing2yaml/feature';
+export type { UsageLimit } from './main/models/pricing2yaml/usage-limit';
+export type { Plan } from './main/models/pricing2yaml/plan';
+export type { AddOn } from './main/models/pricing2yaml/addon';
+export type { RenderMode } from './main/models/pricing2yaml/types';
