+++ src/main/models/pricing2yaml/addon.ts
@@ -1,4 +1,4 @@
-import { valueToNumber } from '../../utils/dzn-exporter/number-utils';
+import { valueToNumber } from '../../../server/utils/dzn-exporter/number-utils';
 import { Feature } from './feature';
 import { UsageLimit } from './usage-limit';
