+++ src/main/utils/pricing-formatting/pricing-parser.ts
