+++ src/main/utils/version-manager.ts
