+++ src/server/models/minizinc/minizinc.ts
