+++ src/server/models/minizinc/raw-models/PricingModel.mzn
