+++ src/server/models/minizinc/raw-models/operations/analysis/cheapest-subscription.mzn
