+++ src/server/models/minizinc/raw-models/operations/analysis/configuration-space.mzn
