+++ src/server/models/minizinc/raw-models/operations/analysis/most-expensive-subscription.mzn
