+++ src/server/models/minizinc/raw-models/operations/filter/cheapest-filtered-subscription.mzn
