+++ src/server/models/minizinc/raw-models/operations/filter/filter.mzn
