+++ src/server/models/minizinc/raw-models/operations/filter/filtered-configuration-space.mzn
