+++ src/server/models/minizinc/raw-models/operations/filter/most-expensive-filtered-subscription.mzn
