+++ src/server/models/minizinc/raw-models/operations/validation/valid-pricing.mzn
