+++ src/types.d.ts
@@ -1,6 +1,6 @@
-export type { Pricing } from './models/pricing2yaml/pricing';
-export type { Feature } from './models/pricing2yaml/feature';
-export type { UsageLimit } from './models/pricing2yaml/usage-limit';
-export type { Plan } from './models/pricing2yaml/plan';
-export type { AddOn } from './models/pricing2yaml/addon';
-export type { RenderMode } from './models/pricing2yaml/types';
+export type { Pricing } from './main/models/pricing2yaml/pricing';
+export type { Feature } from './main/models/pricing2yaml/feature';
+export type { UsageLimit } from './main/models/pricing2yaml/usage-limit';
+export type { Plan } from './main/models/pricing2yaml/plan';
+export type { AddOn } from './main/models/pricing2yaml/addon';
+export type { RenderMode } from './main/models/pricing2yaml/types';
