+++ tests/pricing-service/pricing-service.test.ts
@@ -1,6 +1,6 @@
-import PricingService from '../../src/services/pricing.service';
-import { retrievePricingFromPath } from '../../src/utils/yaml-utils';
-import { Pricing } from '../../src/models/pricing2yaml/pricing';
+import PricingService from '../../src/server/services/pricing.service';
+import { retrievePricingFromPath } from '../../src/main/utils/yaml-utils';
+import { Pricing } from '../../src/main/models/pricing2yaml/pricing';
 import assert from 'assert';
 import { readCSVFile, parseCSVContent } from '../utils/csv-utils';
 import { v4 as uuidv4 } from 'uuid';
